# faster ascii check