--- a/include/simdasciicheck.h
+++ b/include/simdasciicheck.h
@@ -8,11 +8,9 @@
 static bool validate_ascii_fast(const char *src, size_t len) {
   size_t i = 0;
   __m128i has_error = _mm_setzero_si128();
-  __m128i zero = _mm_setzero_si128();
   for (; i + 15 < len; i += 16) {
     __m128i current_bytes = _mm_loadu_si128((const __m128i *)(src + i));
-    has_error =
-      _mm_or_si128(has_error, _mm_cmpgt_epi8(zero,current_bytes));
+    has_error = _mm_or_si128(has_error, current_bytes);
   }
 
   // last part
@@ -21,9 +19,11 @@ static bool validate_ascii_fast(const char *src, size_t len) {
     memset(buffer, 0, 16);
     memcpy(buffer, src + i, len - i);
     __m128i current_bytes = _mm_loadu_si128((const __m128i *)(buffer));
-    has_error =
-      _mm_or_si128(has_error, _mm_cmpgt_epi8(zero,current_bytes));
+    has_error = _mm_or_si128(has_error, current_bytes);
   }
+
+  __m128i signbitmask = _mm_set1_epi8(0b10000000);
+  has_error = _mm_and_si128(has_error, signbitmask);
 
   return _mm_testz_si128(has_error, has_error);
 }