with 4096, 50k iterations, Skylake-EP

```
# before:
validate_ascii_fast(data, N)   :  0.102 cycles per operation (best)    0.118 cycles per operation (avg)
# after:
validate_ascii_fast(data, N)   :  0.086 cycles per operation (best)    0.101 cycles per operation (avg) 
```

I think this is correct... unless a zero byte is considered not valid ASCII.

---

Also, rdtsc counts reference cycles, not core clock cycles, so it's not really a correct benchmark. For comparing two implementations it's okayish (assuming they don't affect the core clock speed differently).