@zbjornson 2018-05-16
Just realized the unit test doesn't test `validate_fast_ascii`, so I haven't actually tested this. Need to run right now, but will come back to this.