@lemire 2018-05-16
The benchmark is correct on a properly configured server. It is definitely not correct on any random machine.

I’ll wait for your test. Thanks.