@zbjornson 2018-05-16
You can easily disable OS-controlled frequency scaling, but I'm not sure that it's possible to totally override the throttling internal to the CPU that's governed by thermal/electrical limits? Maybe you could peg the CPU to some low frequency (here just LFM, but AVX2 base freq if you make a 256-bit version) and correct for the difference between that true frequency, and the base frequency that rdtsc counts at?

---

Code was fine, added a test.