@lemire 2018-05-16
Great code.