@lemire 2018-05-16
Name added to AUTHORS.