@lemire 2018-05-16
I also expect Linux counters to have more overhead. So I would only use them for large tasks.