@cemeyer 2018-05-16
> On my test machine, Linux counters agree exactly (less than 1% error) with rdtsc on sizeable arrays. See 654e660

That is probably true of the SSE implementation today.  It may not continue to be true if you move to higher-power AVX/AVX2/AVX-512 instructions.  The CPU can and does downclock to meet thermal limits under high AVX load (["Workloads using Intel AVX instructions may reduce processor frequency as far down as the AVX base frequency to stay within TDP limits"](http://www.intel.com/content/dam/www/public/us/en/documents/white-papers/performance-xeon-e5-v3-advanced-vector-extensions-paper.pdf)).

I think @zbjornson's concern has merit, for benchmarking an eventual AVX implementation.