@lemire 2018-05-16
@cemeyer 

Fair point, but some things to take into account...

* Intel's documentation is crap and often obsolete or wrong. So I go by what people (ideally me) can measure.
* I have never been able to measure any slowdown with AVX.
* As my blog post [By how much does AVX-512 slow down your CPU? A first experiment](https://lemire.me/blog/2018/04/19/by-how-much-does-avx-512-slow-down-your-cpu-a-first-experiment/) indicates, I have not even been able to measure it with AVX-512.


So though I agree with everything you guys wrote, I would urge some caution. Can you measure these slowdowns on benchmarking servers?

Note that if someone wants to collaborate with me on this, I can make available servers (by ssh).