@zbjornson 2018-05-16
@pcordes do you want to post your measured frequency asm test on SKL-EP with AVX and AVX-512, or mind if I post it?