@pcordes 2018-05-17
@lemire your blog post says you tested on an [Intel Xeon W-2104 CPU @ 3.20GHz](https://ark.intel.com/products/125039/Intel-Xeon-W-2104-Processor-8_25M-Cache-3_20-GHz).

That CPU is special for 3 reasons:

* It *doesn't support turbo* so max freq = baseline all the time.
* It only has one 512-bit FMA unit.
* It's a quad-core (maybe based on the same ring-bus interconnect as regular Skylake-client)?

Using the port5 FMA may be what triggers throttling, so CPUs without the port5 FMA unit may never do the full AVX512 throttling.  (IDK if anyone's done better experiments to figure out exactly what triggers throttling).

AVX512 throttling only seems to go down to the rated baseline sticker clock speed, i.e. 3.2GHz for your CPU.  The higher clocks are for the common case where there's thermal / power headroom, not a guarantee, I don't think.

I'm not totally clear on this point, but I *think* that AVX512 can only slow your CPU down to the same minimum frequency as hitting TDP limits from running heavy 256-bit FMA code on all cores, and that frequency is the sticker frequency that a CPU advertises itself as.  i.e. the sticker frequency is a guaranteed baseline unless your CPU seriously overheats from inadequate cooling, and a different kind of throttling kicks in.

There should be a Stack Overflow chat archive where Mysticial was telling me about overclocking his i9-79xx SKX desktop.  It didn't turn up in some quick googling, but I do remember he said it had a configurable AVX512 frequency in the BIOS, but I don't remember if there was also an AVX frequency below the max turbo for workloads with no FMA or no AVX at all.