@lemire 2018-05-17
@pcordes Note that I always configure testing servers so that they have turbo disabled.