@pcordes 2018-05-18
> Note that I always configure testing servers so that they have turbo disabled.

That's useful for microbenchmarking, especially if you want to use RDTSC instead of perf counters, and what you want to look at for tuning / profiling.

But in real life, people run code with turbo and power-saving enabled, so it's also useful to know how fast different implementations run in wall-clock time.  e.g. in software that doesn't *otherwise* use AVX512 for anything, it's useful to know how an AVX512 implementation compares against an AVX2 implementation, at their respective max clock speeds.

Or with all cores loaded at whatever clock speed the CPU manages...

  This is the kind of thing people need to take into account when deciding which version to use in *their* program on their server.  Although really that's so specific that most people should just benchmark themselves, because the % of time that different programs spend in these functions will be very different.  But it would still be interesting to see how AVX2 vs. AVX512 stacks up on the same hardware, without any artificial clock-speed limitations for either.