@lemire 2018-05-18
@pcordes 

I agree with you. What I want to stress is that one needs to measure these things. Otherwise, we risk taking decisions not based on engineering facts but rather based on stuff we read on the Internet ("AVX-512 slows down your server"). That's no good. Also, Intel's documentation should not be trusted. One needs to measure and verify.