@lemire 2018-05-18
@pcordes I see you live in Halifax. I was once a professor at Acadia University nearby, and I still have friends in Halifax. Nova Scotia is the one place in the world where "brown or white bread?" does not mean what I've come to expect.